+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* In determine_mime_type plugin raise error when file command wasn't found or errored (janko-m)
+
 * Add processing_handler plugin for simpler and more declarative definition of processing (janko-m)
 
 * Storage classes don't need to implement the `#read` method anymore (janko-m)